refactor(granularities): fold granularityConfigHash into GlobalGranularitiesConfig
